examples/aip_dex/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
sqlalchemy_utils==0.41.2
psycopg2-binary==2.9.9
alembic==1.12.1
openai==1.3.5
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.0
pandas==2.1.4
numpy==1.25.2
